fix paging issue

+++ inc/memory.hpp
@@ -2,9 +2,9 @@
 #define DUCK_MEMORY_H
 
 namespace Memory {
-	const uint32_t PAGE_SIZE = 4096;  // 4KiB
-	const uint32_t HUGE_PAGE_SIZE = 2097152;  // 2MiB
-	const uint32_t P3_PAGE_SIZE = 1 << 30;  // 1GiB
+	const uint64_t PAGE_SIZE = 4096;  // 4KiB
+	const uint64_t HUGE_PAGE_SIZE = 2097152;  // 2MiB
+	const uint64_t P3_PAGE_SIZE = 1 << 30;  // 1GiB
 	
 	void init();
